Restructure packaging for pay easy
COAND-497
