+++ payeasy/src/main/java/com/adyen/checkout/payeasy/PayEasyComponent.kt
@@ -16,6 +16,7 @@ import com.adyen.checkout.components.model.payments.request.PayEasyPaymentMethod
 import com.adyen.checkout.components.util.PaymentMethodTypes
 import com.adyen.checkout.econtext.EContextComponent
 import com.adyen.checkout.econtext.EContextDelegate
+import com.adyen.checkout.payeasy.internal.provider.PayEasyComponentProvider
 
 class PayEasyComponent internal constructor(
     delegate: EContextDelegate<PayEasyPaymentMethod>,
